+++ src/main/java/mjuphotolab/photolabbe/auth/jwt/filter/JwtAuthenticationProcessingFilter.java
@@ -1,4 +1,4 @@
-package mjuphotolab.photolabbe.common.jwt.filter;
+package mjuphotolab.photolabbe.auth.jwt.filter;
 
 import java.io.IOException;
 
@@ -7,7 +7,6 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.mapping.GrantedAuthoritiesMapper;
 import org.springframework.security.core.authority.mapping.NullAuthoritiesMapper;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import jakarta.servlet.FilterChain;
@@ -16,8 +15,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import mjuphotolab.photolabbe.common.jwt.service.JwtService;
-import mjuphotolab.photolabbe.common.jwt.util.PasswordUtil;
+import mjuphotolab.photolabbe.auth.CustomUserDetails;
+import mjuphotolab.photolabbe.auth.jwt.service.JwtService;
+import mjuphotolab.photolabbe.auth.jwt.util.PasswordUtil;
 import mjuphotolab.photolabbe.domain.user.entity.User;
 import mjuphotolab.photolabbe.domain.user.repository.UserRepository;
 
@@ -37,7 +37,8 @@ import mjuphotolab.photolabbe.domain.user.repository.UserRepository;
 @RequiredArgsConstructor
 @Slf4j
 public class JwtAuthenticationProcessingFilter extends OncePerRequestFilter {
-	private static final String NO_CHECK_URL = "/login"; // "/login"으로 들어오는 요청은 Filter 작동 X
+	private static final String NO_CHECK_URL_LOGIN = "/api/auth/sign-in";
+	private static final String NO_CHECK_URL_HOME = "/api";
 
 	private final JwtService jwtService;
 	private final UserRepository userRepository;
@@ -47,8 +48,12 @@ public class JwtAuthenticationProcessingFilter extends OncePerRequestFilter {
 	@Override
 	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
 		FilterChain filterChain) throws ServletException, IOException {
-		if (request.getRequestURI().equals(NO_CHECK_URL)) {
-			filterChain.doFilter(request, response); // "/login" 요청이 들어오면, 다음 필터 호출
+
+		/**
+		 * 로그인 여부를 판단하지 않고 진입할 url
+		 */
+		if (request.getRequestURI().equals(NO_CHECK_URL_LOGIN) || request.getRequestURI().equals(NO_CHECK_URL_HOME)) {
+			filterChain.doFilter(request, response);
 			return; // return으로 이후 현재 필터 진행 막기 (안해주면 아래로 내려가서 계속 필터 진행시킴)
 		}
 
@@ -139,22 +144,21 @@ public class JwtAuthenticationProcessingFilter extends OncePerRequestFilter {
 	 * SecurityContextHolder.getContext()로 SecurityContext를 꺼낸 후,
 	 * setAuthentication()을 이용하여 위에서 만든 Authentication 객체에 대한 인증 허가 처리
 	 */
-	public void saveAuthentication(User myUser) {
+	private void saveAuthentication(User myUser) {
+		log.info("saveAuthentication 진입");
+		log.info(myUser.getRole().name());
 		String password = myUser.getPassword();
 		if (password == null) { // 소셜 로그인 유저의 비밀번호 임의로 설정 하여 소셜 로그인 유저도 인증 되도록 설정
 			password = PasswordUtil.generateRandomPassword();
 		}
 
-		UserDetails userDetailsUser = org.springframework.security.core.userdetails.User.builder()
-			.username(myUser.getEmail())
-			.password(password)
-			.roles(myUser.getRole().name())
-			.build();
+		CustomUserDetails userDetailsUser = new CustomUserDetails(myUser);
 
 		Authentication authentication =
 			new UsernamePasswordAuthenticationToken(userDetailsUser, null,
 				authoritiesMapper.mapAuthorities(userDetailsUser.getAuthorities()));
 
 		SecurityContextHolder.getContext().setAuthentication(authentication);
+		log.info("saveAuthentication 끝");
 	}
 }
