+++ src/main/java/mjuphotolab/photolabbe/auth/login/service/LoginService.java
@@ -1,4 +1,4 @@
-package mjuphotolab.photolabbe.common.login.service;
+package mjuphotolab.photolabbe.auth.login.service;
 
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
