+++ src/main/java/mjuphotolab/photolabbe/auth/ouath2/handler/OAuth2LoginSuccessHandler.java
@@ -1,4 +1,4 @@
-package mjuphotolab.photolabbe.common.ouath2.handler;
+package mjuphotolab.photolabbe.auth.ouath2.handler;
 
 import java.io.IOException;
 
@@ -12,8 +12,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import mjuphotolab.photolabbe.common.jwt.service.JwtService;
-import mjuphotolab.photolabbe.common.ouath2.CustomOAuth2User;
+import mjuphotolab.photolabbe.auth.jwt.service.JwtService;
+import mjuphotolab.photolabbe.auth.ouath2.CustomOAuth2User;
 import mjuphotolab.photolabbe.domain.user.entity.Role;
 import mjuphotolab.photolabbe.domain.user.entity.User;
 import mjuphotolab.photolabbe.domain.user.repository.UserRepository;
@@ -38,7 +38,7 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
 			if (oAuth2User.getRole() == Role.GUEST) {
 				String accessToken = jwtService.createAccessToken(oAuth2User.getEmail());
 				response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
-				response.sendRedirect("api/oauth2/extra-sign-up"); // 프론트의 회원가입 추가 정보 입력 폼으로 리다이렉트
+				response.sendRedirect("api/users/mypage"); // 프론트의 회원가입 추가 정보 입력 폼으로 리다이렉트
 
 				jwtService.sendAccessAndRefreshToken(response, accessToken, null);
 				               User findUser = userRepository.findByEmail(oAuth2User.getEmail())
